edk2-ubuntu: set user to be generic
There's no reason to keep the username the same here. As long as the
user id and group id match, the files should be shared properly during
the bind mount. Also, this helps keep things clear when you're in the
container, in case you're like me and you constantly forget.

Signed-off-by: Stephano Cetola <stephanoc@gmail.com>